requirements.txt
pandas
numpy
requests
beautifulsoup4
ipykernel
xgboost
streamlit
python-dotenv
langchain_groq
plotly